QDoc: Include project name in warning and error messages
When building documentation for multiple Qt modules, warnings and errors
don't indicate which module is responsible for issuing them. This makes
it difficult to identify the source of issues, especially when warnings
appear to be about code from a different module than the one being built.

Include the project name in the message prefix to clearly show which
module's documentation build produced each warning or error. This changes
the format from "(qdoc) warning:" to "[ProjectName] (qdoc) warning:".

Fixes: QTBUG-141740
Change-Id: I462fc18115f857b6c98cc7b06f9a2dcab603a67b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>