linguist examples: remove i18n example
This example demonstrates that translations can be provided
later without any modifications to the code. In order to do so, it
provides a dialog with choices for several languages. Upon clicking
on each language, the translation is installed on the app and a
new MainWindow is created to show the new installed translation.
Clicking on more languages creates multiple windows.
The goal of this example is not justified by its implementation.
The documentation is missing and the fact that translations can be
added later doesn't need a whole new example.
A somewhat existing similar example is localizedclock-switchlocale
which demonstrates how the language can be changed at runtime, by
selecting different languages on a list, without creating a new
window everytime.

Task-number: QTBUG-71856
Change-Id: Ib80e4811e166df75cc89cd50fd8edf9213029c69
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>