ltext2id: fix coverity scan INVALIDATE_ITERATOR warning
Coverity complained that transformedItr could be incremented past the
end of the range if the while-loop had already advanced it to end().
It is correct, because the for-loop iterates over 'messages' while
the while-loop advances 'transformedItr' over 'verifyMessages' (two
independent ranges). When a non-supported message was encountered, the
'transformedItr' iterator was incremented unconditionally, but the
while-loop in a previous iteration could have already advanced it
to the end position.

Fix by checking for end() before incrementing. This is the correct
thing to do because the subsequent end-check intentionally doesn't
break out of the loop. It records an error and continues to
accumulate all missing translations. The fix preserves this behavior
and avoids incrementing an iterator that is already advanced to the
end position.

Amends ccddb1221fd05e300e3b5d1ab4ffb5fba49f17db

Coverity-Id: 897288
Change-Id: I7013b664c97fd2cf07f426b53fcfea5518893d8f
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>