QDoc: Respect \inherits command when parsing \qmltype from a .qml file
The QML parser in QDoc always set the base type for a \qmltype as
detected from the type declaration in a .qml file. An \inherits
command was processed but its argument ended up being silently
ignored.

Do not overwrite the argument passed to \inherits. This allows authors
to declare a base type explicitly for documentation purposes.

Fixes: QTBUG-141954
Pick-to: 6.10
Change-Id: I56b474f726d901be357087005ea095bbe349303d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>