QDoc: Enable explicit linking to QML properties and attached properties
Since QML types can have attached properties whose names shadow those of
regular properties, there sometimes needs to be a way to choose which
type of property to link to.

This introduces the "attached" link criterion that enables writers to
express a preference for attached properties while keeping the syntax
minimal for linking to regular properties or uniquely-named attached
properties.

Fixes: QTBUG-142150
Change-Id: If8d45bbcc064c9290bb51110dbbe6c96aa6af399
Reviewed-by: Topi Reinio <topi.reinio@qt.io>