Assistant: Add workaround for styled SVG files in Qt documentation
SVG files in Qt documentation define a <style> element containing
selectors that enable styles for light and dark themes. Online Qt
documentation uses JavaScript to select these but there isn't a simple
alternative to this in Assistant or an obvious way to access the
underlying SVG handling provided by QtSvg.

This workaround adds a "dark" class to the <svg> element when a dark
theme is being used, relying on the SVG file to define appropriate style
rules that can be applied when the class is present.

Change-Id: I84ef450415d2757ac17d4a672b4386d44989082a
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>