QDoc: Adjust order of overloads to more closely match summaries
The order of overloaded functions in detailed sections tends to differ
from that in the lists of functions on some API reference pages. This
can lead to confusion when navigating between overloads or searching for
a suitable overload.

This change imposes some additional order on sorting of functions, based
on the overload number. It does not guarantee that the order will
exactly match the summary list, but it will be closer to it.

Change-Id: I499ac8a22168d252cbaf6a5d8eb5e76fe7090807
Reviewed-by: Topi Reinio <topi.reinio@qt.io>