QDoc: Be more strict about following the inclusion policy
Fix the InclusionFilter::isIncluded() rule to require an exact
match against the Node context flags. This is to eliminate
incorrect behavior when e.g. the policy includes flags for private
functions + showinternal, but the node context is private variable
+ showinternal. Previously, isIncluded() incorrectly returned true
for this case because of the partial match for showinternal.

Follow the inclusion policy in places where it previously
wasn't: replace Node::isInternal() checks with a policy check, or
remove them entirely if appropriate.

In particular, make Node::isInAPI() respect the inclusion policy.
Note that isInAPI() still requires a Node to be documented to
return true; however, generators can (and already do) make
decisions about whether to generate documentation based on the
inclusion policy alone.

Above changes allow now to always set Node status to Node::Internal
when processing the \internal command, regardless of policy, as
the rest of the code follows it as well. The internal status is
visible in the generated .index files. Having the correct status
information allows generators to output e.g. style-related markup
specifically for internal documentation.

Remove a check against declaration location ending in "_p.h" from
ClassNode::docMustBeGenerated() as unnecessary. This choice is now
governed by the `internalfilepatterns` configuration variable.

Simplify docMustBeGenerated() implementations for HeaderNode and
NamespaceNode, and amend their documentation to mention the
inclusion policy.

Pick-to: 6.11
Change-Id: Ic420002e971e27f36a8e81f4b322c0cec3d2ff40
Reviewed-by: Paul Wicking <paul.wicking@qt.io>