QDoc: Issue warnings instead of fatal errors when @-file is invalid
If an include file is not specified, or does not correspond to a valid
file, QDoc issues a fatal warning, causing a core dump. This isn't very
useful for debugging and is inconsistent with other error reporting.

Instead, we now issue a warning and skip the invalid command line
argument.

Fixes: QTBUG-142742
Pick-to: 6.8 6.10 6.11
Change-Id: I27d1725d1ea20e51e1f5fefac1bf15d2be277d8a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>