QDoc: Move \title parsing to DocParser
Remove \title as a meta-command, and instead parse its argument inline
using DocParser, similar to \section argument, as a single-line
paragraph. Store the resulting Text instance to DocPrivate, so that
generators can query a title Text from a Doc and output a formatted
title.

The benefit of this change is that formatting commands are now allowed
within a \title argument. While typical formatting such as \c (code)
or \e (emphasis) are discouraged in titles, \notranslate is useful in
marking terms within titles to exclude them from machine translation.

Keep Node::title() and Node::setTitle(), continue to use them for
title strings such as in .index files, and for the purposes of
searching for nodes using a title. Store a simplified string
representation of the \title Text, with formatting removed, into the
Node for this purpose.

[ChangeLog][QDoc] Formatting commands such as \notranslate are now
supported within page \title argument.

Pick-to: 6.11
Fixes: QTBUG-142530
Change-Id: I1d09db869cd9546eec86ef416ec3315e0dd66754
Reviewed-by: David Boddie <david.boddie@qt.io>