CMake: Add macOS entitlements file for Designer app bundle
The official prebuilt Designer macOS app is codesign-ed with the
hardened runtime option in our post-CI release process, starting with
some patch version of Qt 6.8.

This prevents the application from loading libraries and plugins that
are unsigned or signed with a different identity than the Qt
application itself.

The Widget Designer is meant to load third-party widget plugins, not
only the ones shipped by Qt.

Add an entitlements file to the Designer bundle Resources directory.

Add the 'com.apple.security.cs.disable-library-validation' entitlement
to it which allows the app to load arbitrary plugins.

We don't sign the app during the build process, but our post-CI
release process has code that looks for an '<app_name>.entitlements'
file in the Resources directory and signs the binary with it.

Signing the app during the build process and not shipping the file
does not work, because the post-CI signing process strips the previous
entitlements.

Pick-to: 6.8 6.10 6.11
Fixes: QTBUG-142577
Change-Id: I2e1e492058e91b1e7adc39619e28f137b25828a7
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>