rcc copy: add a missing std::move()
Coverity complained that the 'compressed' variable could have been
moved into 'data', but was instead copied.

It is correct. Add the std::move(), which doesn't impair readability,
because it appears near the end of the scope in which 'compressed' is
defined.

Same issue as CID-464400 in the qtbase copy.

Amends the copying of the code in
950809eb2410d3c17ac5a59535dce07d58016d2a.

Coverity-Id: 426469
Coverity-Id: 464400
Pick-to: 6.10 6.8
Change-Id: Ie22787c034b00379d5bd1825ddea6986219bfeb6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 3db10a1ff0cef3282b706b23601090d87a5da7a0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>