rcc copy: add another missing std::move()
Coverity complained that the 'pwd' variable could have been moved into
the third 'interpretResourceFile' parameter, but was instead copied.

It is correct. Add the std::move(), which doesn't impair readability,
because it appears near the end of the scope in which 'pwd' is defined.

Amends the copying of the code in
950809eb2410d3c17ac5a59535dce07d58016d2a.

Coverity-Id: 427453
Pick-to: 6.10 6.8
Change-Id: Icca84b87724ab52da3214f8c740601149cbbc5a1
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 9a43a856c59c74cfa61fdec408229775bfe4298c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>